TIP-841: Move the channel normalizers to Akeneo\Channel
