+++ src/Akeneo/Channel/Component/spec/Normalizer/ExternalApi/ChannelNormalizerSpec.php
@@ -1,9 +1,9 @@
 <?php
 
-namespace spec\Akeneo\Tool\Component\Api\Normalizer;
+namespace spec\Akeneo\Channel\Component\Normalizer\ExternalApi;
 
 use PhpSpec\ObjectBehavior;
-use Akeneo\Tool\Component\Api\Normalizer\ChannelNormalizer;
+use Akeneo\Channel\Component\Normalizer\ExternalApi\ChannelNormalizer;
 use Akeneo\Channel\Component\Model\ChannelInterface;
 use Prophecy\Argument;
 use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
