+++ src/Akeneo/Channel/Component/spec/Normalizer/InternalApi/ChannelNormalizerSpec.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace spec\Pim\Bundle\EnrichBundle\Normalizer;
+namespace spec\Akeneo\Channel\Component\Normalizer\InternalApi;
 
 use PhpSpec\ObjectBehavior;
 use Pim\Bundle\CatalogBundle\Filter\CollectionFilterInterface;
