+++ src/Pim/Bundle/EnrichBundle/Resources/config/normalizers.yml
@@ -16,7 +16,6 @@ parameters:
     pim_enrich.normalizer.attribute.class:                         Pim\Bundle\EnrichBundle\Normalizer\AttributeNormalizer
     pim_enrich.normalizer.attribute.versioned.class:               Pim\Bundle\EnrichBundle\Normalizer\VersionedAttributeNormalizer
     pim_enrich.normalizer.locale.class:                            Pim\Bundle\EnrichBundle\Normalizer\LocaleNormalizer
-    pim_enrich.normalizer.channel.class:                           Pim\Bundle\EnrichBundle\Normalizer\ChannelNormalizer
     pim_enrich.normalizer.group.class:                             Pim\Bundle\EnrichBundle\Normalizer\GroupNormalizer
     pim_enrich.normalizer.file.class:                              Pim\Bundle\EnrichBundle\Normalizer\FileNormalizer
     pim_enrich.normalizer.product_violation.class:                 Pim\Bundle\EnrichBundle\Normalizer\ProductViolationNormalizer
@@ -174,17 +173,6 @@ services:
         class: '%pim_enrich.normalizer.locale.class%'
         arguments:
             - '@pim_user.context.user'
-        tags:
-            - { name: pim_internal_api_serializer.normalizer }
-
-    pim_enrich.normalizer.channel:
-        class: '%pim_enrich.normalizer.channel.class%'
-        arguments:
-            - '@pim_serializer'
-            - '@pim_enrich.normalizer.locale'
-            - '@pim_versioning.manager.version'
-            - '@pim_enrich.normalizer.version'
-            - '@pim_catalog.filter.chained'
         tags:
             - { name: pim_internal_api_serializer.normalizer }
